removed commented code, utc timehandlng, expiry calc
Time from yahoo is in utc format so translation from string to epoch
and back needs to take this into account. Using datetime.timestamp()
uses the local time while datetime.utcfromtimestamp() uses UTC

Also updated the time to expiration calculation to include the hours
on the day of expiration between midnight and 17:30 which is the
typical time when market closes. Otherwise no option data is reported if
there was less than a full day before the day of expiration.